+++ packages/ag-charts-community/src/chart/axis/axisLabel.ts
@@ -1,4 +1,5 @@
-import type { AgAxisLabelFormatterParams, FontStyle, FontWeight } from '../../options/agChartOptions';
+import type { AgAxisLabelFormatterParams } from '../../options/chart/axisOptions';
+import type { FontStyle, FontWeight } from '../../options/chart/types';
 import { getFont } from '../../scene/shape/text';
 import { Default } from '../../util/default';
 import {
@@ -81,7 +82,7 @@ export class AxisLabel implements ChartAxisLabel {
     avoidCollisions: boolean = true;
 
     /**
-     * By default labels and ticks are positioned to the left of the axis line.
+     * By default, labels and ticks are positioned to the left of the axis line.
      * `true` positions the labels to the right of the axis line.
      * However, if the axis is rotated, it's easier to think in terms
      * of this side or the opposite side, rather than left and right.
