+++ packages/ag-charts-community/src/chart/axis/axisTitle.ts
@@ -1,10 +1,5 @@
-import type {
-    AgAxisCaptionFormatterParams,
-    AgAxisCaptionOptions,
-    FontStyle,
-    FontWeight,
-    TextWrap,
-} from '../../options/agChartOptions';
+import type { AgAxisCaptionFormatterParams, AgAxisCaptionOptions } from '../../options/chart/axisOptions';
+import type { FontStyle, FontWeight, TextWrap } from '../../options/chart/types';
 import {
     BOOLEAN,
     NUMBER,
@@ -31,10 +26,10 @@ export class AxisTitle implements AgAxisCaptionOptions {
     spacing?: number = Caption.PADDING;
 
     @Validate(OPT_FONT_STYLE)
-    fontStyle: FontStyle | undefined = undefined;
+    fontStyle?: FontStyle = undefined;
 
     @Validate(OPT_FONT_WEIGHT)
-    fontWeight: FontWeight | undefined = undefined;
+    fontWeight?: FontWeight = undefined;
 
     @Validate(NUMBER(0))
     fontSize: number = 10;
