+++ packages/ag-charts-community/src/chart/cartesianChart.ts
@@ -1,4 +1,4 @@
-import type { AgCartesianAxisPosition } from '../options/agChartOptions';
+import type { AgCartesianAxisPosition } from '../options/series/cartesian/cartesianOptions';
 import type { BBox } from '../scene/bbox';
 import { toRadians } from '../util/angle';
 import { Logger } from '../util/logger';
