+++ packages/ag-charts-community/src/chart/gridLayout.ts
@@ -1,4 +1,4 @@
-import type { AgChartLegendOrientation } from '../options/agChartOptions';
+import type { AgChartLegendOrientation } from '../options/chart/legendOptions';
 import type { BBox } from '../scene/bbox';
 
 export type Page = { columns: Column[]; pageWidth: number; pageHeight: number; startIndex: number; endIndex: number };
