+++ packages/ag-charts-community/src/chart/series/cartesian/lineSeries.ts
@@ -1,12 +1,9 @@
 import type { ModuleContext } from '../../../module/moduleContext';
-import type {
-    AgCartesianSeriesLabelFormatterParams,
-    AgCartesianSeriesMarkerFormat,
-    AgCartesianSeriesTooltipRendererParams,
-    AgTooltipRendererResult,
-    FontStyle,
-    FontWeight,
-} from '../../../options/agChartOptions';
+import type { AgTooltipRendererResult } from '../../../options/chart/tooltipOptions';
+import type { FontStyle, FontWeight } from '../../../options/chart/types';
+import type { AgCartesianSeriesLabelFormatterParams } from '../../../options/series/cartesian/cartesianLabelOptions';
+import type { AgCartesianSeriesMarkerFormat } from '../../../options/series/cartesian/cartesianSeriesMarkerOptions';
+import type { AgCartesianSeriesTooltipRendererParams } from '../../../options/series/cartesian/cartesianSeriesTooltipOptions';
 import { ContinuousScale } from '../../../scale/continuousScale';
 import { Group } from '../../../scene/group';
 import { PointerEvents } from '../../../scene/node';
