+++ packages/ag-charts-community/src/chart/themes/constants.ts
@@ -1,11 +1,7 @@
-import type {
-    AgCartesianAxisPosition,
-    AgCartesianAxisType,
-    AgChartLegendPosition,
-    AgPolarAxisShape,
-    AgPolarAxisType,
-    FontWeight,
-} from '../../options/agChartOptions';
+import type { AgChartLegendPosition } from '../../options/chart/legendOptions';
+import type { AgPolarAxisShape, AgPolarAxisType } from '../../options/chart/polarAxisOptions';
+import type { FontWeight } from '../../options/chart/types';
+import type { AgCartesianAxisPosition, AgCartesianAxisType } from '../../options/series/cartesian/cartesianOptions';
 
 export const BOLD: FontWeight = 'bold';
 export const NORMAL: FontWeight = 'normal';
