+++ packages/ag-charts-community/src/chart/themes/darkTheme.ts
@@ -1,4 +1,5 @@
-import type { AgChartThemeOptions, AgChartThemePalette } from '../../options/agChartOptions';
+import type { AgChartThemeOptions } from '../../options/agChartOptions';
+import type { AgChartThemePalette } from '../../options/chart/themeOptions';
 import { ChartTheme } from './chartTheme';
 import {
     DEFAULT_AXIS_GRID_COLOUR,
