+++ packages/ag-charts-community/src/module/moduleContext.ts
@@ -10,7 +10,7 @@ import type { LayoutService } from '../chart/layout/layoutService';
 import type { SeriesLayerManager } from '../chart/series/seriesLayerManager';
 import type { SeriesStateManager } from '../chart/series/seriesStateManager';
 import type { UpdateService } from '../chart/updateService';
-import type { AgCartesianAxisPosition } from '../options/agChartOptions';
+import type { AgCartesianAxisPosition } from '../options/series/cartesian/cartesianOptions';
 import type { Scene } from '../scene/scene';
 import type { CallbackCache } from '../util/callbackCache';
